Updated BusStationModel to use LONG for Id and BusLineId properties instead of INT

+++ app/src/main/java/com/example/ratbvkotlin/App.kt
@@ -87,7 +87,7 @@ class App : Application() {
         viewModel { (
                     directionLinkNormal: String,
                     directionLinkReverse: String,
-                    busLineId: Int, busLineName: String
+                    busLineId: Long, busLineName: String
                 ) ->
             BusStationsViewModel(
                 get(),
